# Conflicts:
#	src/App.jsx
#	src/layouts/admin/AdminDashboardLayout.jsx

+++ src/layouts/admin/AdminDashboardLayout.jsx
@@ -43,8 +43,13 @@ export default function AdminDashboardLayout() {
     const handleLogout = async () => {
         const response = await signout();
         if (response && response.status === 200) {
-            localStorage.clear();
-            enqueueSnackbar(response.data.message, {variant: "success", autoHideDuration: 1000});
+            if (localStorage.length > 0){
+                localStorage.clear();
+            }
+            if(sessionStorage.length > 0){
+                sessionStorage.clear()
+            }
+            enqueueSnackbar(response.data.message, { variant: "success", autoHideDuration: 1000 });
             setTimeout(() => (window.location.href = "/home"), 1000);
         }
     };
